fix: export SearchItem and type suggestions to resolve type errors
